requirements.txt
# Основные зависимости для тестирования API мемов

# Фреймворк для тестирования
pytest>=9.0.1

# Библиотека для HTTP запросов
requests>=2.32.5

# Интеграция Allure с pytest для генерации отчетов
allure-pytest>=2.15.2

# Интеграция Allure
allure-python-commons~=2.15.2